refactor(contracts): explicit static module aliasing + canonical imports (#6563)
Replace the dynamic ``sys.modules`` aliasing loop in
``src/shared/python/contracts.py`` (which blindly iterated alias names,
included a self-referential canonical entry, and left ``del`` artifacts)
with two explicit, documented static registrations of the *legacy* alias
names only. Module identity across the three reachable dotted names is
preserved so exception-class identity (e.g. ``PreconditionError``) still
holds for ``except``/``pytest.raises`` across import boundaries.

Migrate the remaining internal importers off the legacy bare ``contracts``
and ``shared.python.contracts`` paths to the canonical
``src.shared.python.contracts`` import path.

``src/shared/python/core/contracts/`` is intentionally a separate concern
(per both module docstrings) and is left untouched.
