+++ src/shared/python/data_processing/tests/test_processor.py
@@ -11,7 +11,7 @@ from unittest.mock import patch
 import numpy as np
 import pandas as pd
 import pytest
-from contracts import PreconditionError
+from src.shared.python.contracts import PreconditionError
 from data_processing.processor import (
     SUPPORTED_FILTER_TYPES,
     DataProcessor,
