+++ src/shared/python/sidekick/process_calculators/pressure_drop_calculator/engine/flow_properties.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import logging
 import math
 
-from shared.python.contracts import require_positive
+from src.shared.python.contracts import require_positive
 
 from ....utils.unit_constants import STANDARD_GRAVITY
 from ...constants import (
